+++ module/tab_warehouse_locations.php
@@ -234,6 +234,10 @@ if ($can_write) {
 		print img_picto('', 'setup', 'class="pictofixedwidth"').$langs->trans('ManageLevels');
 		print '</a>';
 	}
+	$export_url = dol_buildpath('/binloc/admin/import_export.php', 1).'?action=export_assign&fk_entrepot='.$id;
+	print ' <a href="'.$export_url.'" class="button">';
+	print img_picto('', 'download', 'class="pictofixedwidth"').$langs->trans('CsvExportAssignments');
+	print '</a>';
 	print '</div>';
 }